Created a new Search results and Meanings components
The Meanings component is used in the Search Results component where this component is being used on the Dictionary Search component.

Within the meanings component we are looping through the definitions for the results we searched.

also cleans up the manifest that was referring to am image file that no longer exists -  causing an error to be generated in the console.